:hammer: createStatementData 함수 외부 파일로 분리
